Adjusted destructible level objects
I was finally able to figure most of the problems. Shooting it with all projectiles will not only destroy it correctly, but the player will not collide and destroy it right away. When the player collides with the object 10 times, it still explodes. I have to debug this still. But progress.